move the form state test to a separate file
